agent-service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
# The Agentic Stack
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.5
langgraph>=0.2.0
langchain-community>=0.2.0
pinecone>=4.0.0
requests>=2.31.0
pydantic>=2.7.0